% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTable.r
\name{addTable<-,MyClass,data.frame-method}
\alias{addTable<-,MyClass,data.frame-method}
\alias{addTable<-}
\title{Attach or detach MyClass's data table}
\usage{
\S4method{addTable}{MyClass,data.frame}(x, ...) <- value
}
\arguments{
\item{x}{A `MyClass`.}

\item{...}{Nothing}

\item{value}{A `data.frame`, `data.table`, or `matrix`.}
}
\value{
A `MyClass` object.
}
\description{
`addTable()` adds a table to a `MyClass` object.
}
